requirements.txt
h5py>=3.7.0
rasterio>=1.3.4
numpy>=1.22.0
matplotlib>=3.6.0
Pillow>=9.3.0
